Refactor gotodefinition tests to share most via a base class.
